chore: initial restructure commit. tests passing
