+++ integration/basic-app-with-nested-next-config/serverless.yml
@@ -5,7 +5,7 @@ provider:
   runtime: nodejs8.10
 
 plugins:
-  localPath: "../../"
+  localPath: "../../packages/next-plugin"
   modules:
     - index
