+++ packages/next-aws-lambda/aws-lambda-compat.js
