+++ packages/next-plugin/classes/NextPage.js
