+++ packages/next-plugin/classes/PluginBuildDir.js
