+++ packages/next-plugin/classes/__tests__/NextPage.test.js
