+++ packages/next-plugin/classes/__tests__/PluginBuildDir.test.js
