+++ packages/next-plugin/index.js
