+++ packages/next-plugin/lib/__tests__/addCustomStackResources.test.js
