+++ packages/next-plugin/lib/__tests__/copyBuildFiles.test.js
