+++ packages/next-plugin/lib/__tests__/displayStackOutput.test.js
