+++ packages/next-plugin/lib/__tests__/getAssetsBucketName.test.js
