+++ packages/next-plugin/lib/__tests__/parseNextConfiguration.test.js
