+++ packages/next-plugin/lib/__tests__/rewritePageHandlers.js
