+++ packages/next-plugin/lib/__tests__/uploadStaticAssets.test.js
