+++ packages/next-plugin/lib/addCustomStackResources.js
