+++ packages/next-plugin/lib/build.js
