+++ packages/next-plugin/lib/compatLayer.js
