+++ packages/next-plugin/lib/copyBuildFiles.js
