+++ packages/next-plugin/lib/displayStackOutput.js
