+++ packages/next-plugin/lib/getAssetsBucketName.js
