+++ packages/next-plugin/lib/parseNextConfiguration.js
