+++ packages/next-plugin/lib/rewritePageHandlers.js
