+++ packages/next-plugin/lib/uploadStaticAssets.js
