+++ packages/next-plugin/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "serverless-nextjs-plugin",
+  "version": "1.8.1",
+  "description": "A serverless plugin for nextjs 8 serverless target",
+  "main": "index.js",
+  "keywords": [
+    "serverless",
+    "nextjs",
+    "lambda",
+    "next",
+    "api gateway"
+  ],
+  "scripts": {
+    "coveralls": "jest --coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
+    "lint": "eslint .",
+    "integration": "jest --config jest.integration.config.json --setupTestFrameworkScriptFile=./jest.integration.setup.js",
+    "test": "jest"
+  },
+  "author": "Daniel Conde Marin <danielconde9@gmail.com>",
+  "license": "ISC",
+  "peerDependencies": {
+    "next": "^8.0.0"
+  },
+  "devDependencies": {},
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/danielcondemarin/serverless-nextjs-plugin.git"
+  },
+  "dependencies": {
+    "@mapbox/s3urls": "^1.5.3",
+    "chalk": "^2.4.2",
+    "debug": "^4.1.1",
+    "fs-extra": "^7.0.1",
+    "js-yaml": "^3.12.1",
+    "klaw": "^3.0.0",
+    "lodash.clonedeep": "^4.5.0",
+    "lodash.merge": "^4.6.1",
+    "mime": "^2.4.0"
+  },
+  "jest": {
+    "clearMocks": true,
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "index.js",
+      "lib/**/*.js",
+      "utils/**/*.js",
+      "classes/**/*.js"
+    ],
+    "coverageDirectory": "<rootDir>/coverage/",
+    "coveragePathIgnorePatterns": [
+      "<rootDir>/utils/yml/cfSchema.js",
+      "<rootDir>/utils/test"
+    ],
+    "modulePathIgnorePatterns": [
+      "<rootDir>/examples/",
+      "<rootDir>/integration"
+    ]
+  },
+  "engines": {
+    "node": ">=8.5.0"
+  }
+}
