+++ packages/next-plugin/resources/api-gw-proxy.yml
