+++ packages/next-plugin/resources/assets-bucket.yml
