+++ packages/next-plugin/utils/test/parsedNextConfigurationFactory.js
